+++ src/ui.js
@@ -28,6 +28,7 @@ module.exports.body = function(theme) {
     "selection.background": theme.default + "80",
     "menu.background": theme.uibackground,
     "menu.border": theme.uiborder,
+    "menu.separatorBackground": theme.default + "30",
     errorForeground: theme.red,
     contrastBorder: theme.primary + "50",
     contrastActiveBorder: theme.transparent
@@ -245,7 +246,7 @@ module.exports.body = function(theme) {
     "terminal.ansiBlue": theme.blue,
     "terminal.ansiMagenta": theme.pink,
     "terminal.ansiCyan": theme.turquoize,
-    "terminal.ansiWhite": theme.default + "cc",
+    "terminal.ansiWhite": theme.default,
     "terminal.ansiBrightBlack": theme.blue,
     "terminal.ansiBrightRed": theme.red,
     "terminal.ansiBrightGreen": theme.green,
@@ -255,7 +256,7 @@ module.exports.body = function(theme) {
     "terminal.ansiBrightCyan": theme.turquoize,
     "terminal.ansiBrightWhite": theme.primaryalt,
     "terminal.background": theme.uibackground,
-    "terminal.foreground": theme.default + "cc",
+    "terminal.foreground": theme.default,
     "terminalCursor.background": theme.yellow,
     "terminalCursor.foreground": theme.yellow
   };